Fix issue with rendering "line" icons
The svg code from remixicon does not use `stroke` for the line icons,
they are just different paths with a fill. The current `svg_line`
component is redunant becuase it sets properties that are not needed by
the svg icons, like `stroke`, `stroke-width` and `fill="none"`. These
properties also mess up the rendering of the line style icons, by making
them appeared as outlined shapes. The fix is to use a single `svg`
component that will select a different underlying path and nothing more.